Move invitation link files to sub directory
